doc: general TODO list
Some of the todos are related to the expected infrastructure to train and test. Since the infrastructure caontains many requirements (caldera, install beats for sensing etc) we should have two paths on each develompent. 1- Use it for general purpose env (eg docker with defenses and he application to attack, this is easier to maintain and explain for external devs) and 2-Using the mitre tools and other to create an automated attack env and rich sensing (harder to maintain, but good for literature and maybe for executing in an order manner other ways of learning because in thhat env we can do the learning from the execution of any kind of attack from the mitre list and not only generate it for 1 attack)