general: test box, discrete etc containers
Useful for testing problems on shapes with libraries. Gym has not problems with any of them but the built-in rl algorithms seems to have problems if they don't have the 'usual' type in similar envs.
This limits the creation of custom environments (we cannot use whatever type we want for obs or for actions.