SimpleEnv: use a simple discrete problem as general benchmark
Expectations: the learning should be extremely fast (hundreds of steps, tenths of episodes should be enough).
Rsults: so far, too many steps to have good results, and even in that case, the success rate is not 100%.

Test a simple problem against different RL algorithms.

The DRL algorithms suffer from learning simple problems in few steps, even with childish problems like this env.

The env is not polished yet. From experiments, check if the out of bounds proble is inherent to the algorithm or it is a problem/bug in the env step() phase.