IMPORTANT: SimpleEnv working version with only 100 iterations training convergence success
This is a very important commit.

This commit demonstrate the problem of using default PPO settings with step penalty. Depending on the step penalty values, the algorithm converges very easy (100 steps for training is enough for good results, while adding "-10" as time penalty requires 100_000 steps for succesful training.
Te commit contains also the PPO example file for tests.

Note that this is a very simple env, but advantages:
- Easier to try things where convergence problems or failed training appear
- Starting point for other custom environments
- test of algorithms when deploying in production and other custom envs have problems