examples: MAJOR, simple tasks to demonstrate that DRL is extremely inneficient
Since slider demonstrated that DQN and PPO need unaeptable number of training steps to get the right strategy, some more examples are added. For most of them, our future algorithm should solve and quickly (converge).

This is a previous step to the real security environment training (see SimpleConbinationEnv that solves the same problem that we will solve later in security)

The current envs in this folder are:
- simple: learn than action left is always success
- slider: learn to go to a given place in a slider
- combination: learn a subset of switches to success
- goLeft: external example. Will be deleted in the futer. Mainly for comparison to spot errors in the creation of our environments.