algorithms: PPO, GENERAL, avoid harcoded step_block size in PPO for small training steps
Very useful on simple envs where the number of train steps should be small and still learn. In the PPO code, this block is harcoded to 2048 and, opposite to DQN source code, the initial timesteps in configuration cannot be set to 200 because the minimun to execute si in blocks odf 2048 steps in the buffer. So for small steps training, change the n_steps block to an appropriate value.