security: change observation damage count from bool count to type of observation
This is not a "generic" change because it depends on the meaning of the observations. The previous approach was more generic because the more numbers, the more the damage. But this will fit better to the current scenario. In the future we should go back to a generic reward/observation structure to allow more scenarios without changes