PPO: IMPORTANT, reduce slot size where policy must be updated
In simulators, this can be a problem because not many data for policy updates. But, we are in the real system where we have other important constraints as time or max number of iterations.

Therefore, we can afford worse policy updates but much more frequently than before 512/16. This improves a lot the convergence.

Tweaking the learning_rate to 0.01 or 0.001 gave good and bad results depending on the experiment because in some cases it learns too quickly a partial good strategy but it does not reach success etc. So for now do not change learning_rate.