security: DQN tuned to gradient 15 instead of 50 to perform similar n_updates than PPO
The PPO algorithm is tuned so we can check in logs the total network updates n_updates.

To make similar execution on DQN we reduce the gradient updates to 15, because experimentally we see it takes similar updates given similar steps for both algorithms and with the rest of tuned hyperparameters.